pyproject.toml
[tool.poetry]
name = "assignments pml"
version = "0.1.0"
description = "PML 2024 assignments at Copenhagen University"
authors = ["Magnus Samuelsen <97634880+MagnusS0@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^2.1.3"
scipy = "^1.14.1"
pandas = "^2.2.3"
torch = "^2.5.1"
torchvision = "^0.20.1"
matplotlib = "^3.9.2"
pyro-ppl = "^1.9.1"
seaborn = "^0.13.2"
arviz = "^0.20.0"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
